Fix example skipping for FC resources (#1045)
This pull request moves away from using the SkipExamples experimental
bridge feature and uses DocsEdit instead.

Prerequisite for
https://github.com/pulumi/pulumi-terraform-bridge/issues/3152.

Notes: 

- We gain a bunch of supporting type descriptions. 
- Unfortunately, we also lose the Example section for `getFunction`, due
to a limitation with DocsEdit. It's possible to wholesale just remove or
edit content, if we'd like to be more targeted here; I am eager to
remove uses of SkipExamples.
- When https://github.com/pulumi/pulumi/issues/13886 is resolved, we can
remove this logic altogether.

- **FC Service and vsFunction: Elide examples section via docsEdit and
remove SkipExamples use**
- **generate SDK**